Move PyGrid app source files
