--- gateway/requirements.txt
@@ -1,19 +1,0 @@
-Click==7.0
-cryptography
-Flask==1.1.1
-eventlet==0.24.1
-gunicorn==19.9.0
-Flask-Cors==3.0.7
-Flask-Executor==0.9.3
-itsdangerous==1.1.0
-Flask-Sockets==0.2.1
-Jinja2==2.10.1
-MarkupSafe==1.1.1
-gevent==1.4.0
-Werkzeug==0.15.3
-requests==2.22.0
-flask_sqlalchemy
-flask_migrate
-flask_testing
-psycopg2-binary
-git+https://github.com/OpenMined/PySyft@master#egg=syft
