+++ requirements.txt
@@ -1,36 +1,20 @@
-aiounittest==1.3.1
-black
-coverage
-eventlet==0.25.0
-Flask
-Flask
+Click==7.0
+cryptography
+Flask==1.1.1
+eventlet==0.24.1
+gunicorn==19.9.0
+Flask-Cors==3.0.7
 Flask-Executor==0.9.3
-flask_cors
-flask_cors
-flask_login
-flask_login
-flask_migrate
-flask_sockets
+itsdangerous==1.1.0
+Flask-Sockets==0.2.1
+Jinja2==2.10.1
+MarkupSafe==1.1.1
+gevent==1.4.0
+Werkzeug==0.15.3
+requests==2.22.0
 flask_sqlalchemy
+flask_migrate
 flask_testing
-gevent
-gevent-websocket
-numpy>=1.14.0
-numpy>=1.14.0
-pbr
-pre-commit
-pre-commit
-pyjwt
-pytest
-pytest
-redis
-requests
-requests
-requests-toolbelt==0.9.1
--e git+https://github.com/OpenMined/PySyft@master#egg=syft
-syft_proto
--e git+https://github.com/OpenMined/syft-proto@master#egg=syft_proto
-websocket_client
-websocket_client
-websockets>=7.0
-websockets>=7.0
+psycopg2-binary
+jwt
+git+https://github.com/OpenMined/PySyft@master#egg=syft
